actionGrammar: move optimizer benchmarks out of jest into standalone scripts
- Move grammarOptimizerBenchmark and grammarOptimizerSyntheticBenchmark
  from test/*.spec.ts into src/bench/ as plain node entry points.
- Extract shared CONFIGS, runBenchmark, timing, and colored-speedup
  helpers into src/bench/benchUtil.ts.
- Color the speedup column with chalk: green when >1.10x, red when
  <0.90x, plain otherwise.
- Add 'bench', 'bench:synthetic', and 'bench:real' npm scripts; exclude
  dist/bench from the published package.
