+++ ts/packages/actionGrammar/package.json
@@ -24,7 +24,8 @@
   },
   "files": [
     "dist",
-    "!dist/test"
+    "!dist/test",
+    "!dist/bench"
   ],
   "scripts": {
     "build": "npm run tsc",
@@ -36,6 +37,9 @@
     "test:integration": "pnpm run jest-esm --testPathPattern=\"grammarGenerator.spec.js\"",
     "test:local": "pnpm run jest-esm --testPathPattern=\".*[.]spec[.]js\"",
     "test:local:debug": "node --inspect-brk --no-warnings --experimental-vm-modules ./node_modules/jest/bin/jest.js --testPathPattern=\".*\\.spec\\.js\" --testPathIgnorePatterns=\"grammarGenerator.spec.js\"",
+    "bench": "npm run bench:synthetic && npm run bench:real",
+    "bench:synthetic": "node ./dist/bench/grammarOptimizerSyntheticBenchmark.js",
+    "bench:real": "node ./dist/bench/grammarOptimizerBenchmark.js",
     "tsc": "tsc -b"
   },
   "dependencies": {
