+++ ts/packages/actionGrammar/src/bench/grammarOptimizerSyntheticBenchmark.ts
@@ -1,0 +1,125 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+/**
+ * Synthetic optimizer benchmark — informational only.
+ *
+ * Constructs grammars whose structure is *designed* to exercise each
+ * optimization in isolation and at varying scale, so the impact of
+ * each pass is visible without the noise of a real agent grammar.
+ *
+ * Three benchmarks are produced:
+ *
+ *   1. Pass-through chain      — N levels of `<L0> = <L1>; <L1> = <L2>; …`
+ *      Targets `inlineSingleAlternatives`.
+ *
+ *   2. Wide common-prefix      — N alternatives that all start with the
+ *      same long literal prefix and diverge in the last token.
+ *      Targets `factorCommonPrefixes`.
+ *
+ *   3. Combined                — Pass-through wrappers around a wide
+ *      common-prefix block.  Targets both passes together.
+ *
+ * Run with: `pnpm run bench:synthetic` (from this package directory).
+ */
+
+import { runBenchmark } from "./benchUtil.js";
+
+// ─── Synthetic grammar builders ────────────────────────────────────────────
+
+/**
+ * Pass-through chain: `<L0> = <L1>; <L1> = <L2>; …; <LN> = "target"`.
+ * Each `<Li>` adds one nested RulesPart with no other semantics —
+ * exactly the shape `inlineSingleAlternatives` collapses.
+ */
+function buildPassthroughChain(depth: number): string {
+    const lines: string[] = [`<Start> = <L0>;`];
+    for (let i = 0; i < depth; i++) {
+        lines.push(`<L${i}> = <L${i + 1}>;`);
+    }
+    lines.push(`<L${depth}> = target word here;`);
+    return lines.join("\n");
+}
+
+/**
+ * Wide common prefix: N alternatives that share the same long literal
+ * prefix and differ only in the last word.
+ *
+ *   <Choice> = perform the action with item one
+ *            | perform the action with item two
+ *            | …
+ */
+function buildWideCommonPrefix(width: number): string {
+    const prefix = "perform the action with item";
+    const alts: string[] = [];
+    for (let i = 0; i < width; i++) {
+        const word = `value${String.fromCharCode(97 + (i % 26))}${Math.floor(
+            i / 26,
+        )}`;
+        alts.push(`${prefix} ${word} -> "${word}"`);
+    }
+    return `<Start> = <Choice>;\n<Choice> = ${alts.join("\n         | ")};`;
+}
+
+/**
+ * Combined pattern: two layers of pass-through wrapping around a wide
+ * common-prefix block — exercises both passes together.
+ */
+function buildCombined(width: number): string {
+    // Rename <Start> → <Inner> in the wide-prefix grammar and wrap it
+    // in a chain of pass-through rules.  Keep every line of the
+    // renamed inner grammar so <Inner> is actually defined.
+    const inner = buildWideCommonPrefix(width).replace("<Start>", "<Inner>");
+    return [`<Start> = <W1>;`, `<W1> = <W2>;`, `<W2> = <Inner>;`, inner].join(
+        "\n",
+    );
+}
+
+function main(): void {
+    runBenchmark(
+        `pass-through chain (depth=8)`,
+        "synthetic.grammar",
+        buildPassthroughChain(8),
+        ["target word here", "miss", "target word", "no match here"],
+    );
+
+    runBenchmark(
+        `wide common prefix (width=20)`,
+        "synthetic.grammar",
+        buildWideCommonPrefix(20),
+        [
+            "perform the action with item valuea0",
+            "perform the action with item valuet0",
+            "perform the action with item nothere",
+            "perform the action with",
+            "noise input that does not match",
+        ],
+    );
+
+    runBenchmark(
+        `wide common prefix (width=50)`,
+        "synthetic.grammar",
+        buildWideCommonPrefix(50),
+        [
+            "perform the action with item valuea0",
+            "perform the action with item valuex0",
+            "perform the action with item valuew1",
+            "perform the action with item nothere",
+            "noise input",
+        ],
+    );
+
+    runBenchmark(
+        `combined (depth=4 wrappers, width=20 prefix)`,
+        "synthetic.grammar",
+        buildCombined(20),
+        [
+            "perform the action with item valuea0",
+            "perform the action with item valuek0",
+            "perform the action with item nothere",
+            "noise",
+        ],
+    );
+}
+
+main();
